backend/requirements.txt
Flask
flask-cors
langchain-google-genai
langchain-core
google-generativeai
tavily-python
dotenv
gunicorn